move audit logs and docs
