--- audit_logs/__init__.py
@@ -1,3 +1,0 @@
-"""
-:copyright: (c) 2014 Building Energy Inc
-"""
