+++ docs/howto/resetting-the-database.txt
