+++ seed/tasks.py
@@ -22,7 +22,7 @@ from django.db.models.loading import get_model
 from django.core.urlresolvers import reverse_lazy
 from celery.task import task, chord
 from celery.utils.log import get_task_logger
-from audit_logs.models import AuditLog
+from seed.audit_logs.models import AuditLog
 from landing.models import SEEDUser as User
 from seed.lib.mcm import cleaners, mapper, reader
 from mcm.data.ESPM import espm as espm_schema
