+++ seed/urls/projects.py
@@ -1,7 +1,7 @@
 """
 :copyright: (c) 2014 Building Energy Inc
 """
-#!/usr/bin/env python
+# !/usr/bin/env python
 # encoding: utf-8
 """
 urls/urls.py
@@ -13,11 +13,7 @@ from django.conf.urls import patterns, url
 
 urlpatterns = patterns(
     'seed.views.projects',
-    url(
-        r'^add_buildings_to_project/$',
-        'add_buildings_to_project',
-        name='add_buildings_to_project'
-    ),
+    url(r'^add_buildings_to_project/$', 'add_buildings_to_project', name='add_buildings_to_project'),
     url(r'^create_project/$', 'create_project', name='create_project'),
     url(r'^delete_project/$', 'delete_project', name='delete_project'),
     url(
@@ -27,24 +23,11 @@ urlpatterns = patterns(
     ),
     url(r'^get_project/$', 'get_project', name='get_project'),
     url(r'^get_projects/$', 'get_projects', name='get_projects'),
-    url(
-        r'^get_projects_count/$',
-        'get_projects_count',
-        name='get_projects_count'
-    ),
+    url(r'^get_projects_count/$', 'get_projects_count', name='get_projects_count'),
     url(r'^move_buildings/$', 'move_buildings', name='move_buildings'),
-    url(
-        r'^remove_buildings_from_project/$',
-        'remove_buildings_from_project',
-        name='remove_buildings_from_project'
-    ),
+    url(r'^remove_buildings_from_project/$', 'remove_buildings_from_project', name='remove_buildings_from_project'),
     url(r'^update_project/$', 'update_project', name='update_project'),
-    url(
-        r'^update_project_building/$',
-        'update_project_building',
-        name='update_project_building'
-    ),
-    # labels
+    url(r'^update_project_building/$', 'update_project_building', name='update_project_building'),
     url(r'^get_labels/$', 'get_labels', name='get_labels'),
     url(r'^add_label/$', 'add_label', name='add_label'),
     url(r'^apply_label/$', 'apply_label', name='apply_label'),
