+++ seed/views/main.py
@@ -16,7 +16,7 @@ from django.core.files.storage import DefaultStorage
 from django.db.models import Q
 from annoying.decorators import render_to, ajax_request
 from seed.lib.mcm import mapper
-from audit_logs.models import AuditLog
+from seed.audit_logs.models import AuditLog
 from data_importer.models import ImportFile, ImportRecord, ROW_DELIMITER
 from seed.tasks import (
     map_data,
